extracts sortProjects to helper file
