+++ packages/insomnia-app/app/ui/redux/selectors.ts
@@ -5,7 +5,8 @@ import { isWorkspaceActivity } from '../../common/constants';
 import * as models from '../../models';
 import { BaseModel } from '../../models';
 import { getStatusCandidates } from '../../models/helpers/get-status-candidates';
-import { DEFAULT_PROJECT_ID, isRemoteProject, sortProjects } from '../../models/project';
+import { sortProjects } from '../../models/helpers/project';
+import { DEFAULT_PROJECT_ID, isRemoteProject } from '../../models/project';
 import { isRequest, Request } from '../../models/request';
 import { isRequestGroup, RequestGroup } from '../../models/request-group';
 import { UnitTestResult } from '../../models/unit-test-result';
